+++ var/httpd/htdocs/js/Core.Agent.Overview.js
@@ -241,10 +241,12 @@ Core.Agent.Overview = (function (TargetNS) {
                                         $TriggerObj
                                             .next('.ColumnSettingsContainer')
                                             .find('select')
-                                            .after('<span class="SelectedValue Hidden"><a href="#">x</a>' + AutoCompleteText + ' (' + AutoCompleteValue + ')</span>')
+                                            .after('<span class="SelectedValue Hidden">' + AutoCompleteText + ' (' + AutoCompleteValue + ')</span>')
                                             .parent()
-                                            .find('.SelectedValue')
-                                            .find('a')
+                                            .find('input[type=text]')
+                                            .after('<a href="#" class="DeleteFilter"><i class="fa fa-trash-o"></i></a>')
+                                            .parent()
+                                            .find('a.DeleteFilter')
                                             .off()
                                             .on('click', function() {
                                                 $(this)